Implement histogram collection class to facilitate easy interaction with histograms

+++ interface/Category.h
@@ -10,10 +10,10 @@ class Category{
     using pos = std::vector<std::string>::size_type;
     public:
         Category(const std::vector < std::vector <std::string> >&);
-        unsigned getIndex(const std::vector<pos>&);
+        size_t getIndex(const std::vector<pos>&);
         const std::vector<std::string>& getCat() const { return cat; }
     private:
-        std::vector< std::string> cat;
+        std::vector< std::string > cat;
         std::vector< pos > ranges;
 };
 #endif
