+++ src/Category.cc
@@ -20,7 +20,7 @@ Category::Category(const std::vector < std::vector <std::string> >& catVec){
     }
 }
 
-unsigned Category::getIndex(const std::vector<pos>& indices){
+size_t Category::getIndex(const std::vector<pos>& indices){
     if(indices.size() != ranges.size()){
         std::cerr << "Given vector of indices has a wrong dimension, returning index 0!" << std::endl;
         return 0;
